projectmanager/requirements.txt
django
django-tailwind
django-crispy-forms
crispy-tailwind
django-allauth
django-filter
django-htmx
django-money